pyproject.toml
[project]
name = "folktales"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-ollama>=1.0.0",
    "loguru>=0.7.3",
    "networkx>=3.6.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "pyvis>=0.3.2",
    "rdflib>=7.5.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]

